// admin_frontend/src/lib/api/models/SKU.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SKU = {
    sku_code: string;
    weight: number;
    price_cents: number;
    discount_cents?: (number | null);
    quantity: number;
    is_active?: (boolean | null);
    is_visible?: (boolean | null);
    is_limited?: (boolean | null);
    id: number;
    product_id: number;
};